enums: back the four-byte enumerations with int
Comparing the shipped binary's Objective-C ivar metadata against our
own build showed seven fields wider in ours than in the original. Six
are enumerations: the four _thema ivars, RBCustomSelectCollectionView's
_customizeType, and AVBus's mStatus all encode as i in the binary and
as q in our build, because their NS_ENUM was backed by NSInteger.

Back RBUserSettingDataTheme, RBCustomizeItemType, and AVBusStatus with
int so each field keeps the width the binary gives it. The enumerators
and every comparison against them are unaffected.

Two rules said opposite things about this and both are amended to
agree. The integer-width rule exempted NS_ENUM backings outright; it now
extends to them. The enumeration rule forbade a raw int backing; it now
asks for the backing that matches the field's width and signedness, and
warns against widening past what the binary stores.

+++ Project/GameSystem/src/Sound/AVBus.h
@@ -12,7 +12,7 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  * @brief The playback state of an @c AVBus voice, returned by @c status.
  */
-typedef NS_ENUM(NSInteger, AVBusStatus) {
+typedef NS_ENUM(int, AVBusStatus) {
     AVBusStatusNone = -1,    /*!< No source is bound. */
     AVBusStatusPrepared = 1, /*!< A source is bound and prepared but not playing. */
     AVBusStatusPlaying = 2,  /*!< The voice is playing. */
